Move GraphQL bindables to dedicated module
Not moving query and mutation. If they were defined in the bindables
module, the resolvers would not be registered because the schema module
imports from bindables but at that the time nothing has ever been
imported from the resolvers module
