+++ back/boxtribute_server/graph_ql/schema.py
@@ -1,8 +1,15 @@
 from ariadne import make_executable_schema, snake_case_fallback_resolvers
 
+from .bindables import (
+    interface_types,
+    mutation_types,
+    object_types,
+    query_types,
+    union_types,
+)
 from .definitions import definitions, query_api_definitions
 from .enums import enum_types
-from .resolvers import interface_types, mutation, object_types, query, union_types
+from .resolvers import mutation, query
 from .scalars import date_scalar, datetime_scalar
 
 full_api_schema = make_executable_schema(
@@ -12,6 +19,8 @@ full_api_schema = make_executable_schema(
         mutation,
         date_scalar,
         datetime_scalar,
+        *query_types,
+        *mutation_types,
         *object_types,
         *enum_types,
         *union_types,
@@ -26,6 +35,7 @@ query_api_schema = make_executable_schema(
         query,
         date_scalar,
         datetime_scalar,
+        *query_types,
         *object_types,
         *enum_types,
         *union_types,
